Add files via upload
For long-lived species, adult survival is the parameter that has the strongest impact on population dynamics. This makes it crucial to estimate for conservation management. The use of non-invasive genetics approaches coupled with capture–recapture methods is being used increasingly to estimate adult survival. Despite being regularly employed in mammals, it has been little used in birds and notably in long-lived territorial species. This study aimed to determine if feathers obtained from eagle nests could be used to estimate the adult survival rate and the costs associated with such an approach. 

We conducted power tests to evaluate sampling designs (number of territories to sample and time interval between successive samplings) to detect yearly survival probability variation over time of a Golden eagle population. To this end, we simulated a virtual population to which we applied different sampling designs in R 4.1.2 (R Core Team, 2021). Specifically, we simulated life histories of individuals (alive or dead in each year of the simulation) using random draws from a Bernoulli distribution. The life histories of 40 females (for 40 territories) were simulated over 30 years for four values of survival probability: 0.85, 0.875, 0.90 and 0.925. These four values represented survival decreases of 10%, 7.5%, 5%, and 2.5% relative to a control value of 0.95 corresponding to adult golden eagle survival in a stable population (Crandall et al. 2019). We conducted these simulations on females only because very few male feathers were found in nests – too few to estimate their survival within a reasonable budget (see “Results”). When an individual died, it was replaced the year of its death by another individual in the same territory. The history of a new individual was then simulated from this year. This situation represented an area totally occupied by territorial pairs with a systematic turnover due to the large availability of floating individuals (individuals without fixed territories) (Hunt 1998).

On these simulated life histories, we applied different sampling designs from 5 to 30 territories for 5 to 20 years. The simulated sampling designs were “fixed” i.e. the same territories were sampled each year. We also simulated “rotating” designs in which sample territories were drawn randomly and without replacement each year. The simulated data was then analyzed using capture–mark–recapture models (Cormack-Jolly-Seber model) in RMARK to estimate the survival probability of individuals (White and Burnham 1999; Laake 2013) with a probability of detection of 0.8 to mimic situations in which monitored eagles’ pairs occasionally skip reproduction and thus may be undetectable some years.

We conducted each simulation scenario 1000 times in order to assess the expected precision of the survival estimate and its power to detect a lower survival probability compared to the reference value. For each scenario, we used the proportion of survival estimates that did not include the reference value of 0.95 within its 95% confidence interval as the power of the tested scenario. With the usual time constraints in conservation context, we opted for an α-level of 0.2 to favor statistical power, even though it might lead to a higher chance of type I error (Yoccoz, 1991). Thus, if the proportion was greater or equal to 80%, then the statistical power was considered good enough.